Fixed issue with dropped telegraf metrics by increasing batch size
Current configuration of telegraf is transferring 1000 metrics every 10 seconds to InfluxDB. Gateways currently generate about 2000 metrics within 10 seconds, so only half of them were sent, while the rest was buffered. This was causing the buffer to fill up within a few minutes and following metrics to be dropped. This change increases metric batch size to 2000, metric buffer size to 20000 and sets flush_jitter to 5 seconds, to be able to cope with the current amount of metrics.